checkout

+++ src/prototypes.h
@@ -444,15 +444,15 @@ struct CPUINFO{
   struct RAD_MPI **Rsendbuffer;
   struct RAD_MPI **Rrecvbuffer;
 
-  int mpiio_grid_offsets;
-  int *mpiio_ncells;
+  unsigned long int mpiio_grid_offsets;
+  unsigned long int *mpiio_ncells;
 
-  int mpiio_part_offsets;
-  int *mpiio_nparts;
+  unsigned long int mpiio_part_offsets;
+  unsigned long int *mpiio_nparts;
 
 #ifdef STARS
-  int mpiio_star_offsets;
-  int *mpiio_nstars;
+  unsigned long int mpiio_star_offsets;
+  unsigned long int *mpiio_nstars;
 #endif // STARS
 
 #ifdef PIC
